pyproject.toml
[tool.poetry]
name = "hedgehog"
version = "0.3.1"
description = "Random and somewhat useful tools"
authors = ["Aksel Sjögren <akselsjogren@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/akselsjogren/hedgehog"
include = ["bash/*"]
exclude = ["hedgehog/tool.py"]

[tool.poetry.dependencies]
python = "^3.9"
termcolor = "^1.1.0"
simple-term-menu = "^0.10.4"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"
toml = "^0.10.2"

[tool.poetry.scripts]
dirstack = "hedgehog.dirstack.main:main_wrap"
git-fixup = "hedgehog.git.commit:fixup_wrap"
git-pv = "hedgehog.git.commit:preview_wrap"
git-cob = "hedgehog.git.branch:cob_wrap"
git-getbranch = "hedgehog.git.branch:getbranch_wrap"
git-rmb = "hedgehog.git.rmbranch:main_wrap"
hedgehog = "hedgehog.help:main"
hhdiff = "hedgehog.diff.__main__:main"
sshansible = "hedgehog.ssh.main:main_wrap"
fzfdirs = "hedgehog.fzfdirs.main:main_wrap"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
